1.1.4
-Abstracted thread-handling
-Added queueing to threads, to safely return output from threaded
kernels
-Abstracted evolutionary cycle to genetix
-Abstracted calculation for last_difference to genetix
